# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DomusUnify API",
    "version": "v1"
  },
  "paths": {
    "/api/v1/activity/recent": {
      "get": {
        "tags": [
          "Activity"
        ],
        "summary": "Obtém as entradas mais recentes do feed de atividade.",
        "description": "Útil para o cartão \"Recent Updates\" no dashboard (ex.: os primeiros 4 updates).",
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "description": "Número de entradas a devolver (por defeito 4).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityEntryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityEntryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityEntryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity": {
      "get": {
        "tags": [
          "Activity"
        ],
        "summary": "Obtém entradas do feed de atividade (paginado).",
        "description": "Útil para \"View All Activity\".",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "Número de registos a saltar (por defeito 0).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "Número de registos a devolver (por defeito 50).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filtro opcional por tipo/área.\r\nValores suportados: `lists`, `budget`, `calendar`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromUtc",
            "in": "query",
            "description": "Filtro opcional por data/hora inicial (UTC).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toUtc",
            "in": "query",
            "description": "Filtro opcional por data/hora final (UTC, exclusivo).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateUtc",
            "in": "query",
            "description": "Filtro opcional por dia inteiro (UTC). Não pode ser usado em conjunto com fromUtc/toUtc.",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-02-04"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityEntryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityEntryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityEntryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra um novo utilizador com os detalhes fornecidos.",
        "description": "O email deve ser único. A palavra-passe é encriptada antes de ser armazenada.",
        "requestBody": {
          "description": "O pedido de registo contendo nome, email e palavra-passe.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica um utilizador existente e retorna um token de acesso.",
        "description": "Verifica as credenciais fornecidas. Se inválidas, retorna erro de não autorizado.",
        "requestBody": {
          "description": "O pedido de login contendo email e palavra-passe.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/google": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login com Google (ID token devolvido pelo Google Identity Services).",
        "requestBody": {
          "description": "Pedido com o ID token devolvido pelo Google.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalLoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalLoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador da entrada.",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "description": "Tipo/identificador da atividade.",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Mensagem legível (sem incluir o nome do ator).",
            "nullable": true
          },
          "actorUserId": {
            "type": "string",
            "description": "Identificador do utilizador que executou a ação.",
            "format": "uuid"
          },
          "actorName": {
            "type": "string",
            "description": "Nome do utilizador que executou a ação.",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "description": "Data/hora (UTC) da entrada.",
            "format": "date-time"
          },
          "listId": {
            "type": "string",
            "description": "Identificador da lista associada (opcional).",
            "format": "uuid",
            "nullable": true
          },
          "entityId": {
            "type": "string",
            "description": "Identificador de entidade relacionada (opcional).",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Resposta com uma entrada do feed de atividade."
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Token JWT de acesso.",
            "nullable": true
          },
          "expiresAtUtc": {
            "type": "string",
            "description": "Data/hora de expiração do token (UTC).",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Resposta de autenticação com token de acesso."
      },
...
      "ExternalLoginRequest": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "ID token (JWT) devolvido pelo provider.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido de login via provider externo (ex.: Google)."
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do utilizador.",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Palavra-passe do utilizador.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido de autenticação (login)."
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do utilizador.",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "Email do utilizador.",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Palavra-passe do utilizador.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido de registo de utilizador."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Insere o token JWT assim: **Bearer {token}**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}